+++ packages/slice-layout/__tests__/serializers-with-style.js
@@ -1,24 +1,11 @@
-import { AppRegistry } from "react-native-web";
 import {
   addSerializers,
   compose,
   enzymeRenderedSerializer,
   flattenStyleTransform,
   minimalWebTransform,
-  rnwTransform,
   stylePrinter
 } from "@times-components/jest-serializer";
-
-const styles = [
-  "borderBottomColor",
-  "borderBottomWidth",
-  "borderStyle",
-  "paddingBottom",
-  "paddingLeft",
-  "paddingRight",
-  "paddingTop",
-  "width"
-];
 
 addSerializers(
   expect,
@@ -26,8 +13,7 @@ addSerializers(
   compose(
     stylePrinter,
     minimalWebTransform,
-    flattenStyleTransform,
-    rnwTransform(AppRegistry, styles)
+    flattenStyleTransform
   )
 );