+++ packages/slice-layout/src/templates/vertical-list/index.js
@@ -1,5 +1,6 @@
 import React from "react";
-import { FlatList, View } from "react-native";
+import { TcView } from "@times-components/utils";
+import FlatList from "flatlist-react";
 import PropTypes from "prop-types";
 import { ItemRowSeparator } from "../shared";
 
@@ -11,17 +12,14 @@ const ListVerticalLayout = ({ style, tiles }) => {
   return (
     <FlatList
       data={tiles}
-      removeClippedSubviews
       style={style}
-      windowSize={3}
       listKey={listKey}
-      initialNumToRender={5}
       keyExtractor={item => item.props.tileName}
       renderItem={({ item, index }) => (
-        <View key={`${item.props.tileName}`}>
+        <TcView key={`${item.props.tileName}`}>
           {item}
           {index !== tiles.length - 1 ? <ItemRowSeparator /> : null}
-        </View>
+        </TcView>
       )}
     />
   );