+++ packages/section/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.42](https://github.com/newsuk/times-components/compare/@times-components/section@1.9.41...@times-components/section@1.9.42) (2022-06-22)
+
+**Note:** Version bump only for package @times-components/section
+
+
+
+
+
 ## [1.9.41](https://github.com/newsuk/times-components/compare/@times-components/section@1.9.40...@times-components/section@1.9.41) (2022-06-21)
 
 **Note:** Version bump only for package @times-components/section